UBERF-5837 Reorganize collaborator structure for adapters support
